requirements.txt
anthropic
google-api-python-client
youtube-transcript-api
sentence-transformers
chromadb
python-dotenv
streamlit
